fix(dashboard): pylon widget insert
